+++ tests/MongoDB.Driver.Tests/Specifications/unified-test-format/UnifiedErrorMatcher.cs
@@ -20,7 +20,7 @@ using FluentAssertions;
 using MongoDB.Bson;
 using MongoDB.Bson.TestHelpers.XunitExtensions;
 
-namespace MongoDB.Driver.Tests.UnifiedTestOperations
+namespace MongoDB.Driver.Tests.Specifications.unified_test_format
 {
     public class UnifiedErrorMatcher
     {
